+++ tests/testthat/test-scale-date.R
@@ -69,7 +69,7 @@ test_that("datetime colour scales work", {
   expect_equal(range(get_layer_data(p)$colour), c("#132B43", "#56B1F7"))
 })
 
-test_that("date(time) scales throw warnings when input is numeric", {
+test_that("date(time) scales throw warnings when input is incorrect", {
   p <- ggplot(data.frame(x = 1, y = 1), aes(x, y)) + geom_point()
 
   expect_warning(
@@ -80,6 +80,21 @@ test_that("date(time) scales throw warnings when input is numeric", {
   expect_warning(
     ggplot_build(p + scale_x_datetime()),
     "The value was converted to a <POSIXt/POSIXct> object."
+  )
+
+  expect_error(
+    ggplot_build(p + scale_x_date(date_breaks = c(11, 12))),
+    "must be a single string"
+  )
+
+  expect_error(
+    ggplot_build(p + scale_x_date(date_minor_breaks = c(11, 12))),
+    "must be a single string"
+  )
+
+  expect_error(
+    ggplot_build(p + scale_x_date(date_labels = c(11, 12))),
+    "must be a single string"
   )
 
 })